chore: add check to prevent releases on certain types of commits
DSY-778
