--- bin/pre-release/release.js
@@ -1,20 +1,0 @@
-#!/usr/bin/env node
-
-const shell = require('shelljs');
-
-const npmToken = process.env.NPM_TOKEN;
-
-if (!npmToken) {
-  console.log('NPM_TOKEN not found.');
-  process.exit(1);
-}
-
-const init = () => {
-  shell.exec('npm run release');
-  shell.exec('git push --tags');
-  shell.exec('echo "//registry.npmjs.org/:_authToken=$NPM_TOKEN" > ~/.npmrc');
-  shell.exec('npm publish');
-  shell.exec('git push');
-};
-
-init();
